# docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Oikonomia API
  version: 1.0.0
  description: |
    **Oikonomia** (Credit Card Analyzer) — Personal finance management platform for Argentine users.

    ## Architecture
    - **Backend**: FastAPI (Python) with PostgreSQL, Redis, and Celery
    - **Frontend**: Vite + React SPA
    - **Integrations**: Telegram Bot, WhatsApp Bot, Gemini Flash (AI), Yahoo Finance, IOL/PPI brokers

    ## Key Features
    - Expense & income tracking with encrypted fields (AES-256)
    - Credit/debit card management with billing cycle tracking
    - Hierarchical category system with auto-categorization
    - Budget management with 50/30/20 rule support
    - Investment portfolio tracking with auto price refresh
    - AI-powered spending analysis (Gemini Flash)
    - Smart CSV/PDF bank statement import
    - Family group sharing
    - Real-time SSE notifications
    - Scheduled/recurring expense automation
    - Telegram & WhatsApp bot integrations
    - Admin panel with impersonation and audit logging

    ## Security
    - JWT authentication (access + refresh tokens)
    - TOTP-based MFA
    - Field-level AES-256 encryption
    - HMAC for searchable encrypted fields
    - Rate limiting on sensitive endpoints
  contact:
    name: Oikonomia Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8001
    description: Local development
  - url: https://platform.oikonomia.ar
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /categories:
    get:
      tags: [categories]
      summary: List categories for current user
      operationId: listCategories
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
    post:
      tags: [categories]
      summary: Create a new category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /categories/suggest:
    post:
      tags: [categories]
      summary: Get LLM-based category suggestion
      operationId: suggestCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, amount]
              properties:
                description:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Suggestion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategorySuggestResponse'
  /categories/{cat_id}:
    parameters:
      - name: cat_id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [categories]
      summary: Update a category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
    delete:
      tags: [categories]
      summary: Delete a category
      operationId: deleteCategory
      responses:
        '200':
          description: Category deleted

  # ─────────────────────────────────────────────────────────────
  # CARD CLOSINGS
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategoryCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string
          default: '#6366f1'
        keywords:
          type: string
          default: ''
        parent_id:
          type: integer
          nullable: true
        budget_group:
          type: string
          enum: [necesidades, gustos, ahorro]
          default: necesidades
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        keywords:
          type: string
        parent_id:
          type: integer
          nullable: true
        user_id:
          type: integer
          nullable: true
        budget_group:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/CategoryResponse'
    CategorySuggestResponse:
      type: object
      nullable: true
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        confidence:
          type: number

    # ─── CARD CLOSINGS ────────────────────────────────────────
...